Add fallback types for Typescript < 4.1
